tricky/requirements.txt
web3==3.16.5
eth-utils==0.8.1
py-solc==2.1.0
